4.21*: bloglist expansion, step9 end [part4d]

+++ part4/blog-list/controllers/blogs.js
@@ -20,13 +20,11 @@ blogsRouter.get('/', async (req, res) => {
 
 blogsRouter.post('/', async (req, res) => {
   const body = req.body
-  //const token = getTokenFrom(req)
   const decodedToken = jwt.verify(req.token, process.env.SECRET)
   if (!req.token || !decodedToken.id) {
     return res.status(401).json({ error: 'token missing or invalid' })
   }
   const user = await User.findById(decodedToken.id)
-
 
   const blog = new Blog({
     title: body.title,
@@ -44,8 +42,19 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id)
-  res.status(204).end()
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' })
+  }
+  //const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(req.params.id)
+
+  if ( blog.user.toString() === decodedToken.id.toString() ) {
+    await Blog.findByIdAndRemove(req.params.id)
+    res.status(204).end()
+  } else {
+    res.status(401).json({ error: 'no promision to delete' })
+  }
 })
 
 blogsRouter.put('/:id', async (req, res) => {
